add django data relationship

+++ django/django_1n_relationship/articles/views.py
@@ -19,7 +19,9 @@ def create(request):
     if request.method == 'POST':
         form = ArticleForm(request.POST) 
         if form.is_valid():
-            article = form.save()
+            article = form.save(commit=False)
+            article.user = request.user
+            article.save()
             return redirect('articles:detail', article.pk)
     else:
         form = ArticleForm()
@@ -44,38 +46,45 @@ def detail(request, pk):
 @login_required
 @require_http_methods(['GET', 'POST'])
 def update(request, pk):
-    article = Article.objects.get(pk=pk)
-    if request.method == 'POST':
-        form = ArticleForm(request.POST, instance=article)
-        if form.is_valid():
-            form.save()
-            return redirect('articles:detail', article.pk)
-    else:
-        form = ArticleForm(instance=article)
-    context = {
-        'form': form,
-        'article': article,
-    }
-    return render(request, 'articles/update.html', context)
+    article = get_object_or_404(Article, pk=pk)
+    if request.user == article.user:
+        if request.method == 'POST':
+            form = ArticleForm(request.POST, instance=article)
+            if form.is_valid():
+                form.save()
+                return redirect('articles:detail', article.pk)
+        else:
+            form = ArticleForm(instance=article)
+        context = {
+            'form': form,
+            'article': article,
+        }
+        return render(request, 'articles/update.html', context)
+    return redirect('articles:index')
 
 
 @require_POST
 def delete(request, pk):
+    article = get_object_or_404(Article, pk=pk)
     if request.user.is_authenticated:
-        article = Article.objects.get(pk=pk)
-        article.delete()
-    return redirect('articles:index')
+        if request.user == article.user:
+            article.delete()
+        return redirect('articles:index')
+    return redirect('articles:detail', article.pk)
 
 
 @require_POST
 def comments_create(request, pk):
     article = Article.objects.get(pk=pk)
-    comment_form = CommentForm(request.POST)
-    if comment_form.is_valid():
-        comment = comment_form.save(commit=False)
-        comment.article = article
-        comment.save()
-        return redirect('articles:detail', article.pk)
+    comment_form = CommentForm()
+    if request.user.is_authenticated:
+        comment_form = CommentForm(request.POST)
+        if comment_form.is_valid():
+            comment = comment_form.save(commit=False)
+            comment.article = article
+            comment.user = request.user
+            comment.save()
+            return redirect('articles:detail', article.pk)
     context = {
         'article': article,
         'comment_form': comment_form,
@@ -85,6 +94,8 @@ def comments_create(request, pk):
 
 @require_POST
 def comments_delete(request, article_pk, comment_pk):
-    comment = Comment.objects.get(pk=comment_pk)
-    comment.delete()
+    if request.user.is_authenticated:
+        comment = Comment.objects.get(pk=comment_pk)
+        if request.user == comment.user:    
+            comment.delete()
     return redirect('articles:detail', article_pk)
